refactor: move `profile` actions from `api` to `profile`
