+++ lib/screens/OnboardingSubmitCode/presenter.js
@@ -3,7 +3,7 @@ import { setVerificationCode } from '@state/actions/forms'
 import { 
   requestVerificationCode,
   submitVerificationCode
-} from '@state/actions/api'
+} from '@state/actions/profile'
 
 export const mapStateToProps = state => {
   const {
