+++ lib/state/actions/profile.js
@@ -1,3 +1,26 @@
-export const VERIFICATION_SUBMIT_CODE_REQUEST = 'VERIFICATION_SUBMIT_CODE_REQUEST'
+export const VERIFICATION_REQUEST_CODE_REQUEST = 'API:VERIFICATION_REQUEST_CODE_REQUEST'
+export const VERIFICATION_REQUEST_CODE_SUCCESS = 'VERIFICATION_REQUEST_CODE_SUCCESS'
+export const VERIFICATION_REQUEST_CODE_FAILURE = 'VERIFICATION_REQUEST_CODE_FAILURE'
+
+export const VERIFICATION_SUBMIT_CODE_REQUEST = 'API:VERIFICATION_SUBMIT_CODE_REQUEST'
 export const VERIFICATION_SUBMIT_CODE_SUCCESS = 'VERIFICATION_SUBMIT_CODE_SUCCESS'
 export const VERIFICATION_SUBMIT_CODE_FAILURE = 'VERIFICATION_SUBMIT_CODE_FAILURE'
+
+export const requestVerificationCode = (countryCode, phoneNumber) => {
+  return {
+    type: VERIFICATION_REQUEST_CODE_REQUEST,
+    payload: {
+      country_code: countryCode,
+      phone_number: phoneNumber
+    }
+  }
+}
+
+export const submitVerificationCode = (code) => {
+  return {
+    type: VERIFICATION_SUBMIT_CODE_REQUEST,
+    payload: {
+      code
+    }
+  }
+}
