+++ lib/state/reducers/profile/__tests__/verification.test.js
@@ -15,13 +15,13 @@ describe('`profile.verification` reducer', () => {
     })
   })
 
-  describe('on action `VERIFICATION_SUBMIT_CODE_REQUEST`', () => {
+  describe('on action `API:VERIFICATION_SUBMIT_CODE_REQUEST`', () => {
     const state = {
       isLoading: false,
       isVerified: false
     }
     const action = {
-      type: 'VERIFICATION_SUBMIT_CODE_REQUEST',
+      type: 'API:VERIFICATION_SUBMIT_CODE_REQUEST',
       payload: { 
         somePayload: 'somePayload' 
       }
