+++ lib/state/reducers/profile/code-request.js
@@ -2,7 +2,7 @@ import {
   VERIFICATION_REQUEST_CODE_REQUEST,
   VERIFICATION_REQUEST_CODE_SUCCESS,
   VERIFICATION_REQUEST_CODE_FAILURE
-} from '@state/actions/api'
+} from '@state/actions/profile'
 
 const initialState = {
   isLoading: false,
